requirements.txt
Django==2.2.17
djangorestframework==3.9.2
psycopg2-binary==2.8.3
python-dotenv==0.8.2
